+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.7.0] - 2022-03-29
+
 ### Fixed
 
 - Fixed image url in the README.